# demo.security/openapi/main.yaml
openapi: 3.0.3
info:
  title: api documentation of api cm.demo.domeni.security.demo
  description: documentation of api cm.demo.domeni.security.demo
  version: 1.0.0
servers:
  - url: 'https'
paths:
  /user:
    post:
      tags:
        - User
      summary: register user
      description: register user
      operationId: registerUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDTO'
        required: true
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: string
                format: uuid
    get:
      tags:
        - User
      summary: get all users
      description: get all users
      operationId: getAllUsers
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/UserDTO'
  /auth/login:
    post:
      tags:
        - User
      summary: login user
      description: login user
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDTO'
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: string
components:
  schemas:
    UserDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userName:
          type: string
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleDTO'
    CreateUserDTO:
          type: object
          properties:
            id:
              type: string
              format: uuid
            password:
              type: string
              description: password of user to create
            userName:
              type: string
            roles:
              type: array
              items:
                $ref: '#/components/schemas/RoleDTO'
    LoginDTO:
          type: object
          properties:
            username:
              type: string
            roles:
              type: array
              items:
                format: string
            password:
              type: string
    RoleDTO:
      properties:
        name:
          type: string